Refactored adding messages to session for clean architecture
