+++ src/main/java/callhub/connect/CallhubconnectApplication.java
@@ -1,13 +1,9 @@
 package callhub.connect;
 
-import callhub.connect.data_access.MessageDataAccess;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -40,11 +36,6 @@ public class CallhubconnectApplication {
 
 			}
 		};
-	}
-
-	@Bean
-	public MessageDataAccess messageDataAccess(ApplicationContext context) {
-		return new MessageDataAccess(context);
 	}
 
 }
