+++ src/main/java/callhub/connect/use_case/message/MessageDataAccessInterface.java
@@ -1,7 +1,11 @@
 package callhub.connect.use_case.message;
 
+import callhub.connect.entities.Sender;
+
 import java.util.HashMap;
 
 public interface MessageDataAccessInterface {
     public HashMap<String, String> generateResponse(String message);
+
+    public void sendResponseToDatabase(String content, String sessionId, Sender sender);
 }
